requirements.txt
pandas
yfinance
matplotlib
pyyaml